+++ resources/js/signerapp/components/Stepper.vue
@@ -1,5 +1,10 @@
 <template>
   <div class="w-100 px-lg-5 px-0">
+    <div class="mb-3">
+      <a href="./">
+        <b-img style="height:40px" fluid src="/images/signfinger.svg"></b-img>
+      </a>
+    </div>
     <b-card v-for="(v,i) in data" :key="i" class="font-weight-bold shadow" :class="[step == i+1 ? 'bg-warning':' bg-c-dark-grey text-white d-none d-lg-block',v.class]" no-body>
       <b-card-body class="py-2">