+++ resources/js/website/localization.js
@@ -1,0 +1,32 @@
+export default {
+  "en.howWork": {
+    "title": "Works as easy as counting to three",
+    "1": {
+      "title": "Attach your file or document",
+      "subtitle": "Select your desired file or document (in PDF format) to be signed and attach it to the browser"
+    },
+    "2": {
+      "title": "Draw a signature on the browser",
+      "subtitle": "Define an area as signature target on the document and draw your signature on the given pad in the platform"
+    },
+    "3": {
+      "title": "Download it",
+      "subtitle": "Preview it before download. If it is not as desired, just repeat it. simple"
+    }
+  },
+  "id.howWork": {
+    "title": "Bekerja semudah menghitung sampai tiga",
+    "1": {
+      "title": "Lampirkan dokumen anda",
+      "subtitle": "Pilih dokumen anda (dalam format PDF) yang akan ditandatangani dan lampirkan pada browser",
+    },
+    "2": {
+      "title": "Gambar tanda tangan pada browser",
+      "subtitle": "Tentukan area yang menjadi tempat tanda tangan anda pada dokumen dan gambar tanda tangan anda di blok yang disediakan pada platform",
+    },
+    "3": {
+      "title": "Unduh",
+      "subtitle": "Tinjau kembali hasil tanda tangan anda sebelum mengunduh. Jika tidak sesuai dengan yang diharapkan, ulangi kembali. Semudah itu",
+    },
+  },
+}