+++ resources/views/scripts/website.blade.php
@@ -1,1 +1,2 @@
 <script src="{{ asset('js/website/app.js') }}"></script>
+<script src="{{ mix('js/website.js') }}"></script>