+++ webpack.mix.js
@@ -30,7 +30,9 @@ mix.webpackConfig({
 });
 mix
   .js('resources/js/signerapp/app.js', 'public/js/signerapp.js')
-  .sass('resources/js/signerapp/custom.scss', 'public/css/style.css');
+  .js('resources/js/website/app.js', 'public/js/website.js')
+  .sass('resources/js/signerapp/custom.scss', 'public/css/style.css')
+  .sass('resources/sass/website.scss', 'public/css/website/style.css');